requirements.txt
numpy==1.26.4
opencv-python
ultralytics
pygame
pyttsx3
google-generativeai
torch
torchvision
torchaudio
requests